Require ToS before we verify
